- getMemberBorrowedBooks method in Service and Controller and updated deleteReviewFromBook method. - Created printBooks method in Controller for cleaner code. - updated isStaff method in Controller

+++ src/LibraryService/LibraryService.java
@@ -93,6 +93,22 @@ public class LibraryService {
         }
     }
 
+
+    /**
+     * Gets all the books a member has borrowed.
+     *
+     * @param memberID the ID of the member
+     * @return a list of borrowed books by the member
+     */
+    public List<Book> getMemberBorrowedBooks(int memberID) {
+        List<Book> books = new ArrayList<>();
+        Member member = memberRepo.get(memberID);
+        for(Loan loan : member.getLoanHistory()) {
+            books.add(loan.getBook());
+        }
+        return books;
+    }
+
     /**
      * Deletes a review from a book.
      *
@@ -103,8 +119,15 @@ public class LibraryService {
         Review reviewToDelete = reviewRepo.get(reviewID);
         if (reviewToDelete != null) {
             Book book = reviewToDelete.getBook();
-            book.getReviews().remove(reviewToDelete);
-            reviewRepo.delete(reviewID);
+            Member member = reviewToDelete.getMember();
+            List<Loan> memberLoans = member.getLoanHistory();
+
+            boolean memberHasBorrowedBook = memberLoans.stream()
+                    .anyMatch(loan -> loan.getBook().getID() == book.getID());
+            if (memberHasBorrowedBook) {
+                book.getReviews().remove(reviewToDelete);
+                reviewRepo.delete(reviewID);
+            }
         }
     }
